requirements.txt
--find-links=http://pypi.pinaxproject.com
--find-links=http://pypi2.pinaxproject.com

django-ajax-validation==0.1.3
django-announcements==0.1.0
django-avatar==1.0.3
django-dbtemplates==0.5.8
django-notification==0.1.4
django-pagination==1.0.5.1
django-robots==0.6.1
django-messages==0.4.2
django-threadedcomments==0.5.1
django-extensions==0.4.1
django-uni-form==0.6.0
django-timezones==0.1.4
django-bookmarks==0.1.0
django-email-confirmation==0.1.4
django-gravatar==0.1.0
django-mailer==0.1.0
django-friends==0.1.4
django-locations==0.1.1
django-oembed==0.1.0
django-swaps==0.5.3
django-flag==0.1.1
django-db-log==1.0.1
django-attachments==0.3
django-markup==0.3
django-microblogging==0.1.2
django-wikiapp==0.2.0
sorl==3.1
django-generic-flatblocks==0.3.0
django-frontendadmin==0.4
django-filter==0.5.1
django-photologue==2.3
django-tagging==0.3

# packages we create ourselves for release:
django-voting==0.1_pre
django-openid
django-sorting>=0.1
template_utils==0.4p2
django-staticfiles==0.1.2
django-debug-toolbar==0.8.0
django-listings==0.1
django-events==0.1
django-app-plugins==0.1.0

